+++ inc/modules/V1785N.h
@@ -321,7 +321,7 @@ namespace vmepp
             void            PrintStatus();
 
             //Data acquisition
-            bool            GetEventAt( uint32_t index, UEvent<V1785N> &event ) const override;
+            bool            GetEventAt( size_t index, UEvent<V1785N> &event ) const override;
             uint32_t        GetBufferAddress() const override { return 0; }
 
             void            ClearData();
